communicator/requirements.txt
paho-mqtt
numpy
keras
datetime
requests